requirements.txt
python-dotenv
requests
discord.py==2.0.1
py-cord